refactor: edit filename in fetch directory
